requirements.txt
Django==3.1.5
psycopg2-binary==2.8.4
djangorestframework
